requirements.txt
# Core ML/DL frameworks
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0

# Web framework
Flask>=2.3.0

# Data processing and visualization
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0
Pillow>=9.0.0
opencv-python>=4.7.0

# Machine learning utilities
scikit-learn>=1.2.0
tqdm>=4.64.0

# Development and monitoring
tensorboard>=2.12.0

# Optional: For enhanced data loading and processing (uncomment if needed)
# datasets>=2.10.0
# transformers>=4.20.0

# Optional: For additional model architectures (uncomment if needed)
# timm>=0.6.0

# Optional: For advanced visualization (uncomment if needed)
# plotly>=5.0.0